Laravel 5.2. CMS

+++ app/Http/Controllers/BlogController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Blog as Blog;
-use App\Region as Region;
 use App\User as User;
 use App\Country as Country;
 use App\Category as Category;
@@ -30,11 +29,10 @@ class BlogController extends Controller
 
     public function create()
     {
-        $regions = Region::all();
         $users = User::all();
         $countries = Country::orderBy('name', 'asc')->get();
         $workflows = Workflow::orderBy('id','desc')->get();
-        $data = ['countries' => $countries,'regions' => $regions,'users' => $users,'workflows' => $workflows];
+        $data = ['countries' => $countries,'users' => $users,'workflows' => $workflows];
         return view('admin.createblog')->with($data);
     }
 
@@ -44,7 +42,7 @@ class BlogController extends Controller
 
         $errors = Validator::make($request->all(), [
             'title' => 'required|max:255',
-            'region_id' => 'required',
+
             'description' => 'required',
             'image' => 'required',
             ]);
@@ -55,7 +53,7 @@ class BlogController extends Controller
             ->withInput();
         }
 
-        $request['region_id'] = implode(",", $request['region_id']);
+
         $request['title'] = strip_tags($request['title']);
         $request['slug'] = str_slug($request['title']);
         $request['subtitle'] = strip_tags($request['subtitle']);
@@ -86,19 +84,19 @@ class BlogController extends Controller
                 $constraint->aspectRatio();
             });
 
-            $watermark = Image::make(public_path() . '/assets/img/watermark.png')->resize(300, null, function ($constraint) {
-                $constraint->aspectRatio();
-            });;
-
-            $watermarkthumb = Image::make(public_path() . '/assets/img/watermark.png')->resize(50, null, function ($constraint) {
-                $constraint->aspectRatio();
-            });
-
-            $watermarkthumb->opacity(30);
-            $watermark->opacity(30);
-
-            $imagethumb->insert($watermarkthumb, 'center');
-            $imagemedium->insert($watermark,'center');
+//            $watermark = Image::make(public_path() . '/assets/img/watermark.png')->resize(300, null, function ($constraint) {
+//                $constraint->aspectRatio();
+//            });;
+//
+//            $watermarkthumb = Image::make(public_path() . '/assets/img/watermark.png')->resize(50, null, function ($constraint) {
+//                $constraint->aspectRatio();
+//            });
+//
+//            $watermarkthumb->opacity(30);
+//            $watermark->opacity(30);
+//
+//            $imagethumb->insert($watermarkthumb, 'center');
+//            $imagemedium->insert($watermark,'center');
 
             $imagethumb->save($pathThumb . $imageName);
             $imagemedium->save($pathMedium . $imageName);
@@ -133,12 +131,11 @@ class BlogController extends Controller
 
     public function edit($id)
     {
-        $regions = Region::all();
         $users = User::all();
         $blog = Blog::FindOrFail($id);
         $countries = Country::orderBy('name', 'asc')->get();
         $workflows = Workflow::orderBy('id','desc')->get();
-        $data = ['countries' => $countries,'regions' => $regions,'users' => $users,'workflows' => $workflows,'blog' => $blog];
+        $data = ['countries' => $countries,'users' => $users,'workflows' => $workflows,'blog' => $blog];
         return view('admin.editblog')->with($data);
     }
 
@@ -147,7 +144,7 @@ class BlogController extends Controller
     {
         $errors = Validator::make($request->all(), [
             'title' => 'required|max:255',
-            'region_id' => 'required',
+
             'description' => 'required',
             ]);
 
@@ -157,7 +154,7 @@ class BlogController extends Controller
             ->withInput();
         }
 
-        $request['region_id'] = implode(",", $request['region_id']);
+
         $request['title'] = strip_tags($request['title']);
         $request['slug'] = str_slug($request['title']);
         $request['subtitle'] = strip_tags($request['subtitle']);
@@ -219,14 +216,17 @@ class BlogController extends Controller
     {
         $blog = Blog::FindOrFail($id);
 
-        // Delete blog images
-        $image = public_path() . '/assets/img/blog/'.$blog->image;
-        $imagemedium = public_path() . '/assets/img/blog/medium/'.$blog->image;
-        $imagethumb = public_path() . '/assets/img/blog/thumbnails/'.$blog->image;
+        if($blog->image) {
+            // Delete blog images
+            $image = public_path() . '/assets/img/blog/'.$blog->image;
+            $imagemedium = public_path() . '/assets/img/blog/medium/'.$blog->image;
+            $imagethumb = public_path() . '/assets/img/blog/thumbnails/'.$blog->image;
 
-        unlink($image);
-        unlink($imagemedium);
-        unlink($imagethumb);
+            unlink($image);
+            unlink($imagemedium);
+            unlink($imagethumb);
+        }
+
 
         $blog->delete();
         return redirect('/admin/blog');
